+++ CHANGELOG.md
@@ -18,5 +18,6 @@ All notable changes to this project are documented in this file.
 - Implement `Contact` library. (https://github.com/dic-iit/bipedal-locomotion-framework/pull/43 and https://github.com/dic-iit/bipedal-locomotion-framework/pull/45)
 - Added `CommonConversions` and `ManifConversions` libraries to handle type conversions.
 - Implement the `JointPositionTracking` application. (https://github.com/dic-iit/bipedal-locomotion-framework/pull/136)
+- Initial implementation of Python bindings using pybind11 (https://github.com/dic-iit/bipedal-locomotion-framework/pull/134)
 
 [Unreleased]: https://github.com/dic-iit/bipedal-locomotion-framework/